Create runtime.txt
When following the instructions to create a new app, I now get "FileNotFoundError: [Errno 2] No such file or directory: '/app/.config/gcloud/application_default_credentials.json'" in the Heroku logs when pushing. Specifying an older version of Python fixes this. Not sure if we want to keep patching like this, but bringing it to your attention.